Create kpbot2.addrenewer.py
This was, until this moment, by far the most complex script/program that I have made.
I had to combine my scrapping skills with my selenium skills, and not only for scrapping some data and pictures but also for some data which was used as part of xpath. I had to learn that it is possible to manipulate xpath. Wonderful.
Also, I managed to learn more on os and win32.com libraries, and how to use, and abuse, windows through python.